refactor(intents): move SiteEntity into AnglesiteIntents lib + accept injectable store (#104)
